Ramp hashes & validations
